add ability to turn off buffer pooling
