+++ geode-core/src/distributedTest/java/org/apache/geode/distributed/internal/P2PMessagingConcurrencyDUnitTest.java
@@ -15,6 +15,9 @@
 
 package org.apache.geode.distributed.internal;
 
+import static org.apache.geode.internal.lang.SystemPropertyHelper.BYTE_BUFFER_POOL_STRATEGY;
+import static org.apache.geode.internal.lang.SystemPropertyHelper.GEODE_PREFIX;
+import static org.apache.geode.internal.lang.SystemPropertyHelper.getProductStringProperty;
 import static org.apache.geode.test.awaitility.GeodeAwaitility.await;
 import static org.assertj.core.api.Assertions.assertThat;
 
@@ -340,6 +343,9 @@ public class P2PMessagingConcurrencyDUnitTest {
      * careful: if you set a boolean it doesn't take hold! setting a String
      */
     props.setProperty("conserve-sockets", String.valueOf(conserveSockets));
+
+//    getProductStringProperty(BYTE_BUFFER_POOL_STRATEGY).map( poolType ->
+//      props.setProperty(GEODE_PREFIX + BYTE_BUFFER_POOL_STRATEGY, poolType));
 
     return props;
   }
