+++ geode-core/src/main/java/org/apache/geode/distributed/internal/direct/DirectChannel.java
@@ -14,6 +14,9 @@
  */
 package org.apache.geode.distributed.internal.direct;
 
+import static org.apache.geode.internal.lang.SystemPropertyHelper.BYTE_BUFFER_POOL_STRATEGY;
+import static org.apache.geode.internal.lang.SystemPropertyHelper.getProductStringProperty;
+
 import java.io.IOException;
 import java.io.NotSerializableException;
 import java.net.InetAddress;
@@ -50,6 +53,8 @@ import org.apache.geode.internal.cache.DirectReplyMessage;
 import org.apache.geode.internal.inet.LocalHostUtil;
 import org.apache.geode.internal.logging.log4j.LogMarker;
 import org.apache.geode.internal.net.BufferPool;
+import org.apache.geode.internal.net.BufferPoolImpl;
+import org.apache.geode.internal.net.BufferPoolNoOp;
 import org.apache.geode.internal.tcp.BaseMsgStreamer;
 import org.apache.geode.internal.tcp.ConnectExceptions;
 import org.apache.geode.internal.tcp.Connection;
@@ -116,7 +121,13 @@ public class DirectChannel {
     this.receiver = listener;
     this.dm = dm;
     this.stats = dm.getStats();
-    this.bufferPool = new BufferPool(stats);
+
+    final String poolType = getProductStringProperty(BYTE_BUFFER_POOL_STRATEGY).orElse("pool");
+    if ("none".equalsIgnoreCase(poolType)) {
+      this.bufferPool = new BufferPoolNoOp();
+    } else {
+      this.bufferPool = new BufferPoolImpl(stats);
+    }
 
     DistributionConfig dc = dm.getConfig();
     this.address = initAddress(dc);
