+++ geode-core/src/test/java/org/apache/geode/internal/net/NioSslEngineTest.java
@@ -82,7 +82,7 @@ public class NioSslEngineTest {
 
     mockStats = mock(DMStats.class);
 
-    final BufferPool bufferPool = new BufferPool(mockStats);
+    final BufferPool bufferPool = new BufferPoolImpl(mockStats);
     spyBufferPool = spy(bufferPool);
     nioSslEngine = new NioSslEngine(mockEngine, spyBufferPool);
     spyNioSslEngine = spy(nioSslEngine);
@@ -125,7 +125,7 @@ public class NioSslEngineTest {
     verify(mockEngine, atLeast(2)).getHandshakeStatus();
     verify(mockEngine, times(3)).wrap(any(ByteBuffer.class), any(ByteBuffer.class));
     verify(mockEngine, times(3)).unwrap(any(ByteBuffer.class), any(ByteBuffer.class));
-    verify(spyBufferPool, times(2)).expandWriteBufferIfNeeded(any(BufferPool.BufferType.class),
+    verify(spyBufferPool, times(2)).expandWriteBufferIfNeeded(any(BufferPoolImpl.BufferType.class),
         any(ByteBuffer.class), any(Integer.class));
     verify(spyNioSslEngine, times(1)).handleBlockingTasks();
     verify(mockChannel, times(3)).read(any(ByteBuffer.class));
@@ -212,7 +212,7 @@ public class NioSslEngineTest {
       try (final ByteBufferSharing outputSharing2 = spyNioSslEngine.wrap(appData)) {
         ByteBuffer wrappedBuffer = outputSharing2.getBuffer();
 
-        verify(spyBufferPool, times(1)).expandWriteBufferIfNeeded(any(BufferPool.BufferType.class),
+        verify(spyBufferPool, times(1)).expandWriteBufferIfNeeded(any(BufferPoolImpl.BufferType.class),
             any(ByteBuffer.class), any(Integer.class));
         appData.flip();
         assertThat(wrappedBuffer).isEqualTo(appData);
@@ -424,14 +424,14 @@ public class NioSslEngineTest {
   public void ensureWrappedCapacityOfSmallMessage() {
     ByteBuffer buffer = ByteBuffer.allocate(netBufferSize);
     assertThat(
-        nioSslEngine.ensureWrappedCapacity(10, buffer, BufferPool.BufferType.UNTRACKED))
+        nioSslEngine.ensureWrappedCapacity(10, buffer, BufferPoolImpl.BufferType.UNTRACKED))
             .isEqualTo(buffer);
   }
 
   @Test
   public void ensureWrappedCapacityWithNoBuffer() {
     assertThat(
-        nioSslEngine.ensureWrappedCapacity(10, null, BufferPool.BufferType.UNTRACKED)
+        nioSslEngine.ensureWrappedCapacity(10, null, BufferPoolImpl.BufferType.UNTRACKED)
             .capacity())
                 .isEqualTo(netBufferSize);
   }
