+++ geode-core/src/test/java/org/apache/geode/internal/tcp/MsgStreamerTest.java
@@ -39,11 +39,12 @@ import org.apache.geode.distributed.internal.DistributionMessage;
 import org.apache.geode.distributed.internal.SerialAckedMessage;
 import org.apache.geode.distributed.internal.membership.InternalDistributedMember;
 import org.apache.geode.internal.net.BufferPool;
+import org.apache.geode.internal.net.BufferPoolImpl;
 import org.apache.geode.internal.serialization.KnownVersion;
 
 public class MsgStreamerTest {
   private DMStats stats = mock(DMStats.class);
-  private BufferPool pool = spy(new BufferPool(stats));
+  private BufferPool pool = spy(new BufferPoolImpl(stats));
   Connection connection1 = mock(Connection.class);
   Connection connection2 = mock(Connection.class);
